index.ts: Add `log` module
The `log` module contains `critical`, `error`, `warning`, `info`
and `debug` functions that each take a format string a la

    "first value = {}, second value = {}"

and a string array with format string arguments. Due to limitations
in AssemblyScript, we can at this point not provide a more convenient
solution that converts arguments to strings behind the scenes.