common: more native types support for BigInt and ByteArray
added BigInt.fromU64, BigInt.fromI64 and BigInt.fromU32 as well as corresponding on ByteArray and `toU64`, `toI64` method methods

Refactored test.js file at the same time to use recent Node.js features to reduce the code and make the asynchronous operation use `async/await`. Error are properly reported in test now with a link to the failed assertion and a message if provided.