Improved the typings of the `assert` function
TypeScript has a nice way to inform the type checker that a function asserts a certain condition and that after the assertion, the rest of the scope must assume the condition is true.

What this entails is that it's now possible to have much more interesting code when it's time to deal with `null` value.


```
  assert(out != null, `JSON key ${key} is not found in JSON Object.`)
  out.toObject().get("other")
```

Here, `out` after the `assert` is known to the type checker to be `non-null` because the condition `out != null` now holds for the whole scope after the `assert`.

The same code before the modification in this PR:

```
  if (out == null) {
    throw `JSON key ${key} is not found in JSON Object.`
  }
  out.toObject().get("other")
```

So we gain clarity.

See https://github.com/AssemblyScript/assemblyscript/issues/2195

#### Breaking Change

It's not technically a breaking change because only the type system is involved.