Added `mustGet` an `mustGetEntry` to TypedMaped
This adds two small helpers that make it easier to work with `JSONValue` so that it's possible to get elements from objects more tersely.

#### Open Questions

I noticed that `Entity` which extends `TypeMap` uses `get` without checking `null` and only using typing tricks to remove the nullability that `get` brings.

So I'm wondering:
- Should I remove the `assert` in the `must...` version
- Or should we change `Entity` to now use `mustGet` instead